Changed time step to be in sec instead of minutes. Also made it a double value.
git-svn-id: file://localhost/data1/git/presto_svn/trunk@634 5fac236c-a0d8-0310-9745-ef116914eda1

+++ include/rfifind_cmd.h
@@ -35,9 +35,9 @@ typedef struct s_Cmdline {
   char rfixwinP;
   /***** -rfips: Plot the RFI instances in a PS file */
   char rfipsP;
-  /***** -time: Minutes to integrate for stats and FFT calcs */
+  /***** -time: Seconds to integrate for stats and FFT calcs */
   char timeP;
-  float time;
+  double time;
   int timeC;
   /***** -timesig: The +/-sigma cutoff to reject time-domain chunks */
   char timesigmaP;
